+++ scrapper/src/main/java/edu/java/dto/response/LinkUpdateResponse.java
@@ -4,7 +4,7 @@ import java.net.URI;
 import java.util.List;
 
 public record LinkUpdateResponse(
-    long id,
+    Integer id,
     URI url,
     String description,
     List<Long> tgChatIds
